requirements.txt
alembic==1.9.1
fastapi==0.88.0
sqlmodel==0.0.8
uvicorn==0.18.2
psycopg2-binary==2.9.5
python-dotenv==0.20.0
spacy==3.4.4
ru-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/ru_core_news_sm-3.4.0/ru_core_news_sm-3.4.0-py3-none-any.whl
beautifulsoup4==4.11.1
python-multipart==0.0.5
tenacity==8.0.1
SQLAlchemy==1.4.41
ruts==0.8.1
spacy-transformers==1.2.0
sentencepiece==0.1.99
protobuf==3.20
